Add tmxlib.image.open(); fix and test add_image_layer
